refactor : MineSweeperConsole.kt 추가, 매직넘버 제거
